pyproject.toml
[project]
authors = [
  {name = "pritam-dey3", email = "pritam.pritamdey.984@gmail.com"},
]
dependencies = [
  "networkx>=3.5",
  "pillow>=11.3.0",
  "pyautogui>=0.9.54",
  "python-statemachine[diagrams]",
  "opencv-python-headless>=4.12.0.88",
  "transitions>=0.9.3",
  "graphviz>=0.21",
]
description = "Advance GUI automation"
name = "pyautoguide"
readme = "README.md"
requires-python = ">=3.13"
version = "0.4.0"

[project.scripts]
pyautoguide = "pyautoguide:main"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.uv.sources]
python-statemachine = {git = "https://github.com/fgmacedo/python-statemachine.git", rev = "9a089edf61b4afd4abc08968b3cfa508fa332f7c"}

[dependency-groups]
dev = [
  "pre-commit>=4.2.0",
  "types-networkx>=3.5.0.20250712",
]

[project.optional-dependencies]
ocr = [
  "rapidocr>=3.2.0",
  "onnxruntime>=1.22.0",
]

[tool.poe.tasks]
precmt = "pre-commit run --all-files"

[tool.ruff.format]
skip-magic-trailing-comma = true

[tool.ruff]
ignore = ["E731"]
line-length = 88
